Build xz for darwin-x86
Add options handling in utility build_package method.

+++ library/utility.rb
@@ -1,3 +1,4 @@
+require 'digest'
 require_relative 'formula.rb'
 require_relative 'platform.rb'
 require_relative 'build.rb'
@@ -43,9 +44,7 @@ class Utility < Formula
   end
 
   def build_package(release, options)
-    # todo:
-    #platforms = options.platforms
-    platforms = [Platform.new('darwin-x86_64')]
+    platforms = options.platforms.map { |name| Platform.new(name) }
     puts "Building #{name} #{release} for platforms: #{platforms.map{|a| a.name}.join(' ')}"
 
     @num_jobs = options.num_jobs
@@ -64,15 +63,18 @@ class Utility < Formula
       @log_file = build_log_file(platform)
       # build
       FileUtils.cd(build_dir) { build_for_platform platform, release, options }
-      # todo: check options
+      next if options.build_only?
       # package
       archive = File.join(Global::CACHE_DIR, archive_filename(release, platform.name))
       FileUtils.rm_f archive
       args = ['-C', base_dir, '-Jcf', archive, ARCHIVE_TOP_DIR]
       Utils.run_command 'tar', *args
-      # todo: update sha sum
-      # todo: install if same platform
       #
+      if options.update_shasum?
+        release.shasum = { platform.to_sym => Digest::SHA256.hexdigest(File.read(archive, mode: "rb")) }
+        update_shasum release, platform
+      end
+      install_archive release, archive if Global::PLATFORM_NAME == platform.name
       FileUtils.rm_rf base_dir unless options.no_clean?
     end
 
@@ -95,8 +97,11 @@ class Utility < Formula
 
   def install_archive(release, archive)
     rel_dir = release_directory(release)
-    FileUtils.rm_rf rel_dir
-    FileUtils.mkdir_p rel_dir
+    # else we'll fail while updating 'xz'
+    if active_version != release.to_s
+      FileUtils.rm_rf rel_dir
+      FileUtils.mkdir_p rel_dir
+    end
     Utils.unpack archive, Global::NDK_DIR
     write_active_file File.basename(rel_dir)
   end
@@ -129,7 +134,42 @@ class Utility < Formula
     File.join base_dir_for_platform(platform), 'build.log'
   end
 
-  # def package_dir
-  #   File.join build_base_dir, 'package'
-  # end
+  def update_shasum(release, platform)
+    ver = release.version
+    cxver = release.crystax_version
+    sum = release.shasum(platform.to_sym)
+    release_regexp = /^[[:space:]]*release[[:space:]]+version:[[:space:]]+'#{ver}',[[:space:]]+crystax_version:[[:space:]]+#{cxver}/
+    platform_regexp = /#{platform.to_sym}:/
+    lines = []
+    state = :copy
+    File.foreach(path) do |l|
+      case state
+      when :updated
+        lines << l
+      when :copy
+        if  l !~ release_regexp
+          lines << l
+        else
+          if l !~ platform_regexp
+            state = :updating
+            lines << l
+          else
+            state = :updated
+            lines << l.gsub(/'[[:xdigit:]]+'/, "'#{sum}'")
+          end
+        end
+      when :updating
+        if l !~ platform_regexp
+          lines << l
+        else
+          state = :updated
+          lines << l.gsub(/'[[:xdigit:]]+'/, "'#{sum}'")
+        end
+      else
+        raise "in formula #{File.basename(formula_file)} bad state #{state} on line: #{l}"
+      end
+    end
+
+    File.open(path, 'w') { |f| f.puts lines }
+  end
 end
